Generate the `src/lib.rs` from template files
Add crate documentation about the device this crate has be build for.

Add some `make` commands to simplify development, as we have to add a
feature to every cargo command line. We can't use the [dev-dependencies]
to store one for use as this make changing the device we want to build
for more difficult. It would have needed to add `--no-default-features
--features=<device>`. The chosen way make it explicit when our run
`cargo` command that you must add a device feature.

The device features added on the `Cargo.toml` file has been done with
the `make prepare` command that use `stm32builder` binary (not the lib
we use on the build script` to update `Cargo.toml` from device files
defined in the `devices/` folder.