RCC: Provides private API for handling peripherals state
The RCC has the responsibility to enable, disable and reset all
peripherals on device. So provide an API for other peripheral to ask it
to do this on there behalf.

The FIXME are here because the `stm32-rs` doesn't defined the needed
values. This will be fixed later.

The reset bit handling is wrong, will be fixed when updating to use
`modify_reg!()` api.

Data comes from `peripherals.gpio.banks[*].bus` defined in device files
that has been mapped to `peripherals.rcc.buses`.